Rename opentracing-spring-web-autoconfigure to *-starter
Fixes: #75
