--- opentracing-spring-web-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.opentracing.contrib.spring.web.autoconfig.ServerTracingAutoConfiguration,\
-io.opentracing.contrib.spring.web.autoconfig.RestTemplateTracingAutoConfiguration
