+++ opentracing-spring-web-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+io.opentracing.contrib.spring.web.starter.ServerTracingAutoConfiguration,\
+io.opentracing.contrib.spring.web.starter.RestTemplateTracingAutoConfiguration
